requirements.txt
attrs
pyparsing